+++ tests/testthat/test-class-worksheet.R
@@ -108,3 +108,25 @@ test_that("print options work", {
   expect_equal(exp, got)
 
 })
+
+test_that("ignore_error works", {
+
+  wb <- wb_workbook()$add_worksheet()
+  wb$add_data(dims = "B1", x = t(c(1, 2, 3)), colNames = FALSE)
+  wb$add_formula(dims = "A1", x = "SUM(B1:C1)")
+
+  # 1
+  wb$worksheets[[1]]$ignore_error(dims = "A1", formulaRange = TRUE)
+
+  exp <- "<ignoredErrors><ignoredError sqref=\"A1\" formulaRange=\"1\"/></ignoredErrors>"
+  got <- wb$worksheets[[1]]$ignoredErrors
+  expect_equal(exp, got)
+
+  # 2
+  wb$worksheets[[1]]$ignore_error(dims = "A2", calculatedColumn = TRUE, emptyCellReference = TRUE, evalError = TRUE, formula = TRUE, formulaRange = TRUE, listDataValidation = TRUE, numberStoredAsText = TRUE, twoDigitTextYear = TRUE, unlockedFormula = TRUE)
+
+  exp <- "<ignoredErrors><ignoredError formulaRange=\"1\" sqref=\"A1\"/><ignoredError formulaRange=\"1\" sqref=\"A2\" calculatedColumn=\"1\" emptyCellReference=\"1\" evalError=\"1\" formula=\"1\" listDataValidation=\"1\" numberStoredAsText=\"1\" twoDigitTextYear=\"1\" unlockedFormula=\"1\"/></ignoredErrors>"
+  got <- wb$worksheets[[1]]$ignoredErrors
+  expect_equal(exp, got)
+
+})